+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - feat: Bank accounts are now the default account type.
 - feat: Quick bank name options for PH banks when creating a bank account.
+- feat: Detect finance app usage with permission toggle.
+- feat: Request notification permission on Notifications screen and recurring transactions.
 
 ## [0.33.0] - 2025-05-20
 ### Added